Fix #39 by ignoring updates to unauthorized fields instead of failing.
Previously, if a user attempted to update fields when commenting on
an issue but the user did not have permission to update the fields the
whole request would be denied. This problematic when the HTML form
did not fully populate all of the fields (like when a bug has an Area
that is no longer in the admin approved list). So rather than fail
outright, we'll now update the fields the user is allowed to update and
ignore the rest.